requirements.txt
openai
datasets
langchain_nvidia_ai_endpoints
langchain_community
faiss-cpu
tenacity
together
python-dotenv
streamlit
sentence-transformers
google-cloud-aiplatform
tqdm
